+++ assets/effects/Stun.py
@@ -1,0 +1,10 @@
+from engine.game.effect.effect import Effect
+
+class Stun(Effect):
+    """Sets target's speed to 0 in order to stun"""
+
+    def __init__(self, duration):
+        super().__init__("stunned", duration)
+
+    def on_build_action(self, action):
+        return 0
